+++ HISTORY.rst
@@ -6,6 +6,7 @@ History
 ------------------
 
 - Handle uncaught errors via custom excepthook (nens/rana-qgis-plugin#254)
+- Modified (user facing) texts (nens/rana#3136)
 
 
 1.2.8 (2026-03-13)